pyproject.toml
[tool.poetry]
name = "overload"
version = "0.1.0"
description = "DoS Tool"
authors = ["7zx, 8fn and João"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
requests = "^2.28.1"
colorama = "^0.4.5"
humanfriendly = "^10.0"
PySocks = "^1.7.1"
scapy = "^2.4.5"
get-mac = "^0.8.3"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
pytest = "^7.1.3"
mypy = "^0.982"
types-requests = "^2.28.11.2"
deptry = "^0.6.4"
prospector = "^1.7.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"